requirements.txt
pycryptodome==3.21.0
